// Docker/Docker_Service_Face/ai-projects/oryza-ai-frontend/constants/enum.ts
export enum ResultEnum {
  success = "SUCCESS",
  error = "ERROR",
}
export enum TypeSocketServer {
  server = "ALIVE_SERVER",
  service = "ALIVE_SERVICE",
}

export enum CameraType {
  NORMAL_CAMERA = 'NORMAL_CAMERA',
  PTZ_CAMERA = 'PTZ_CAMERA',
  AI_CAMERA = 'AI_CAMERA',
  TRAFFIC_CAMERA = 'TRAFFIC_CAMERA',
}